+++ index.html
@@ -12,8 +12,8 @@
     <header class="container">
       <h1><a href="/">영화 리스트</a></h1>
 
-      <form id="search">
-        <input id="searchInput" type="text" placeholder="검색어를 입력해주세요." />
+      <form id="search" action="search.html" method="get">
+        <input id="searchInput" type="text" name="query" placeholder="검색어를 입력해주세요." />
         <button id="searchBtn" type="submit">검색</button>
       </form>
     </header>
