+++ src/movie.js
@@ -66,3 +66,12 @@ export async function loadPost({ id, backdrop_path, title, release_date, genres,
   let genreIds = genres.map((a) => a.id);
   createGenreList(id, genreIds);
 }
+
+export function checkLastPage() {
+  const cardCount = document.querySelectorAll("#movies > div").length;
+  const morePageBtn = document.getElementById("morePage");
+  if (cardCount < 20) {
+    morePageBtn.disabled = true;
+    morePageBtn.textContent = cardCount === 0 ? "검색결과가 없습니다." : "다음 페이지가 없습니다.";
+  }
+}
