+++ src/search.js
@@ -1,13 +1,11 @@
 // 검색 페이지 파일
 import { searchPage, loadGenre } from "./fetch.js";
 import { handleClickCard } from "./movie.js";
-import { navigateToSearch } from "./utils.js";
 
 let currentPage = 1;
 
-const keyword = new URL(location.href).searchParams.get("query");
-
 async function load() {
+  const keyword = new URL(location.href).searchParams.get("query");
   await loadGenre();
   searchPage(currentPage, keyword);
 }
@@ -21,4 +19,3 @@ load();
 document.getElementById("searchInput").focus();
 document.getElementById("morePage").addEventListener("click", morePage);
 document.getElementById("movies").addEventListener("click", handleClickCard);
-document.getElementById("search").addEventListener("submit", navigateToSearch);
